requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
ultralytics==8.4.14
pillow==12.1.1
torch==2.11.0+cu128
torchvision==0.26.0+cu128
pylabel==0.1.55
kagglehub==1.0.0
fiftyone==1.13.2
tqdm==4.67.3
numpy==2.4.2
opencv-python==4.13.0.92
diffusers
accelerate
transformers
roboflow==1.3.1
python-dotenv==1.2.2
shapely==2.1.2